Feat/tgym 008

+++ src/main/java/com/project/gymapp/modules/products/services/ProductService.java
@@ -2,14 +2,18 @@ package com.project.gymapp.modules.products.services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.project.gymapp.modules.products.exceptions.ProductAlreadyRegisteredWithIdException;
+import com.project.gymapp.modules.products.exceptions.ProductByTypeNotFoundException;
+import com.project.gymapp.modules.products.exceptions.ProductNotFoundException;
+import com.project.gymapp.modules.products.exceptions.ProductsNotFoundException;
 import com.project.gymapp.modules.products.models.Product;
 import com.project.gymapp.modules.products.models.dtos.ProductDTO;
 import com.project.gymapp.modules.products.repositories.ProductRepository;
-import com.project.gymapp.modules.utils.CreateUUID;
 
 @Service
 public class ProductService {
@@ -17,20 +21,23 @@ public class ProductService {
     @Autowired
     ProductRepository productRepository;
 
-    @Autowired
-    CreateUUID createUUID;
-
-    public ProductService(ProductRepository productRepository, CreateUUID createUUID) {
+    public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
     }
 
     public List<Product> getAllProducts() {
         List<Product> products = productRepository.findAll();
+        if (products.isEmpty()) {
+            throw new ProductsNotFoundException();
+        }
         return products;
     }
 
     public List<Product> getAllProductsByProductType(String productType) {
         List<Product> productsByProductType = productRepository.findByProductType(productType);
+        if (productsByProductType.isEmpty()) {
+            throw new ProductByTypeNotFoundException();
+        }
         return productsByProductType;
     }
 
@@ -42,10 +49,10 @@ public class ProductService {
         }
 
         if (productFounded.isPresent()) {
-            throw new Exception("Produto já cadastrado no sistema!");
+            throw new ProductAlreadyRegisteredWithIdException();
         }
 
-        String id = createUUID.createUUID();
+        String id = UUID.randomUUID().toString();
         Product product = new Product(id, productDTO.brand(), productDTO.manufacturingDate(), productDTO.name(), productDTO.price(), productDTO.quantity(), productDTO.productDetails(), productDTO.productType(), productDTO.valiDate());
         Product productToSave = productRepository.save(product);
         return productToSave;
@@ -53,11 +60,17 @@ public class ProductService {
 
     public Optional<Product> findById(String id) {
         Optional<Product> product = productRepository.findById(id);
+        if (product.isEmpty()) {
+            throw new ProductNotFoundException();
+        }
         return product;
     }
 
     public void deleteById(String id) {
         Optional<Product> product = productRepository.findById(id);
+        if (product.isEmpty()) {
+            throw new ProductNotFoundException();
+        }
         Product productToDelete = product.get();
         productRepository.delete(productToDelete);
     }
@@ -65,36 +78,40 @@ public class ProductService {
     public Product updateProduct(ProductDTO productDTO, String id) {
         Optional<Product> product = productRepository.findById(id);
 
+        if (product.isEmpty()) {
+            throw new ProductNotFoundException();
+        }
+
         if (product.isPresent()) {
-            if (productDTO.name() != null && productDTO.name().equals(product.get().getName())) {
+            if (productDTO.name() != null && !productDTO.name().equals(product.get().getName())) {
                 product.get().setName(productDTO.name());
             }
 
-            if (productDTO.brand() != null && productDTO.brand().equals(product.get().getBrand())) {
+            if (productDTO.brand() != null && !productDTO.brand().equals(product.get().getBrand())) {
                 product.get().setBrand(productDTO.brand());
             }
 
-            if (productDTO.price() != null && productDTO.price().equals(product.get().getPrice())) {
+            if (productDTO.price() != null && !productDTO.price().equals(product.get().getPrice())) {
                 product.get().setPrice(productDTO.price());
             }
 
-            if (productDTO.manufacturingDate() != null && productDTO.manufacturingDate().equals(product.get().getManufacturingDate())) {
+            if (productDTO.manufacturingDate() != null && !productDTO.manufacturingDate().equals(product.get().getManufacturingDate())) {
                 product.get().setManufacturingDate(productDTO.manufacturingDate());
             }
 
-            if (productDTO.valiDate() != null && productDTO.valiDate().equals(product.get().getValiDate())) {
+            if (productDTO.valiDate() != null && !productDTO.valiDate().equals(product.get().getValiDate())) {
                 product.get().setValiDate(productDTO.valiDate());
             }
 
-            if (productDTO.productDetails() != null && productDTO.productDetails().equals(product.get().getProductDetails())) {
+            if (productDTO.productDetails() != null && !productDTO.productDetails().equals(product.get().getProductDetails())) {
                 product.get().setProductDetails(productDTO.productDetails());
             }
 
-            if (productDTO.productType() != null && productDTO.productType().equals(product.get().getProductType())) {
+            if (productDTO.productType() != null && !productDTO.productType().equals(product.get().getProductType())) {
                 product.get().setProductType(productDTO.productType());
             }
 
-            if (productDTO.quantity() != null && productDTO.quantity().equals(product.get().getQuantity())) {
+            if (productDTO.quantity() != null && !productDTO.quantity().equals(product.get().getQuantity())) {
                 product.get().setQuantity(productDTO.quantity());
             }
